Pipeline: Add pipeline data types, define TimeBuffer Factory instance.

+++ src/common/types.hh
@@ -12,10 +12,10 @@ typedef uint16_t PhyReg; // Physical reg name type
 
 typedef uint32_t Imm; // immediate data type
 
-typedef uint8_t Xlen; // Arch X length type
+typedef uint64_t Xlen; // Arch X length type
 
 typedef uint64_t Reg; // Reg data type
 
 typedef uint16_t ThreadId; // Thread ID type
 
-typedef uint64_t Latency;
+typedef uint64_t Latency; // Stage latency type
